Refactor Bernoulli and FirstLinear to Classes
