+++ sympy/solvers/ode/tests/test_single.py
@@ -4,7 +4,7 @@
 #
 
 from sympy.core import Function, Symbol
-from sympy.solvers.ode.single import (NthAlgebraic, ODEMatchError,
+from sympy.solvers.ode.single import (FirstLinear, ODEMatchError,
     SingleODEProblem, SingleODESolver)
 
 from sympy.testing.pytest import raises
@@ -24,16 +24,16 @@ def test_SingleODESolver():
     raises(NotImplementedError, lambda: solver._matches())
     raises(NotImplementedError, lambda: solver._get_general_solution())
 
-    # This ODE can not be solved by the NthAlgebraic solver. Here we test that
+    # This ODE can not be solved by the FirstLinear solver. Here we test that
     # it does not match and the asking for a general solution gives
     # ODEMatchError
 
-    problem = SingleODEProblem(f(x).diff(x) + f(x), f(x), x)
+    problem = SingleODEProblem(f(x).diff(x) + f(x)*f(x), f(x), x)
 
-    solver = NthAlgebraic(problem)
+    solver = FirstLinear(problem)
     raises(ODEMatchError, lambda: solver.get_general_solution())
 
-    solver = NthAlgebraic(problem)
+    solver = FirstLinear(problem)
     assert solver.matches() is False
 
     #These are just test for order of ODE
